Lint cleanup: remove dead null-handling and redundant idioms
K2's stronger flow analysis flags several genuinely-dead constructs:
- PreferenceItems: '!!' on onClick, smart-cast non-null under 'if (clickable)'
- TripExtrapolationBuilder: 'state?.' where 'projection != null' already implies
  state non-null (projection = state?.let{})
- Open311Screen: '?.' on decodeSampledBitmapFromFile's non-null Bitmap return
- ScheduleReplayExtrapolator: '?: return null' on non-null schedule.stopTimes
- TripResultsRepository: redundant .toLong() on an already-Long duration
- DirectionsGenerator: unreachable 'else -> null' in two exhaustive enum whens
- SurveyUtils: two dead 'currentStop.id == null' checks (id is a non-null type)

All behavior-preserving; the removed branches never executed. The remaining
icon-when 'else -> {log; -1}' is left in place as a deliberate caller-facing
sentinel.

+++ onebusaway-android/src/main/java/org/onebusaway/android/ui/tripresults/TripResultsRepository.kt
@@ -51,7 +51,7 @@ class DefaultTripResultsRepository @Inject constructor(@param:ApplicationContext
     ): Result<List<ItineraryOption>> = withContext(Dispatchers.IO) {
         runCatching {
             itineraries.map { itinerary ->
-                val durationSec = itinerary.duration.toLong()
+                val durationSec = itinerary.duration
                 ItineraryOption(
                     title = DirectionsGenerator(itinerary.legs, context).itineraryTitle,
                     durationText = ConversionUtils
